Safari-611.1.10/LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-hide-esc-key-expected.txt

Testing a TracksPanel is hidden upon hitting the Esc. key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


mediaControls.showTracksPanel()

Pressing the Escape key

Transition ended
PASS mediaControls.tracksPanel.presented is false

PASS successfullyParsed is true

TEST COMPLETE
